requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
networkx>=3.2
pyvis>=0.3.2
requests>=2.31.0
python-multipart>=0.0.9
httpx>=0.27.0
feedparser>=6.0.11
pypdf2>=3.0.1
python-docx>=1.1.0
reportlab>=4.1.0
